statsd: Increase max packet size to 8k
The MTU inside our network is large enough to support transmiting large
UDP packets; the largest the packet, the more metrics we can fit in
multi-metric mode and the more throughput for the daemon.

The receive buffer has been switched to heap-allocated, because 8k may
be slightly too much for the stack.